+++ subjects/PE/x86/pySample/pySample_text.c
@@ -36,7 +36,7 @@ PyObject * fn100010F0(PyObject * ptrArg04, PyObject * ptrArg08)
 {
 	PyObject * eax_n = PyArg_ParseTuple(ptrArg08, "ff:fdiv", fp - 0x08, fp - 0x04);
 	if (eax_n != null)
-		return Py_BuildValue("f", (real64) rLoc08 / rLoc04);
+		return Py_BuildValue("f", (real64) rLoc08 / (real64) rLoc04);
 	return eax_n;
 }
